refactor: embed restrictions into transfer namespace
