+++ src/entities/SecurityToken/Transfers/index.ts
@@ -1,0 +1,1 @@
+export { Transfers } from './Transfers';
